+++ services/todo/internal/handler/handler_list.go
@@ -6,8 +6,8 @@ import (
 	"errors"
 
 	"github.com/brunoluiz/go-lab/core/app"
-	"github.com/brunoluiz/go-lab/services/apollo/gen/openapi"
-	"github.com/brunoluiz/go-lab/services/apollo/internal/repo"
+	"github.com/brunoluiz/go-lab/services/todo/gen/openapi"
+	"github.com/brunoluiz/go-lab/services/todo/internal/repo"
 	"github.com/segmentio/ksuid"
 )
